+++ R/CRSSIO.R
@@ -1,7 +1,6 @@
 #' @keywords internal
 "_PACKAGE"
 
-#' @importFrom magrittr "%>%"
 #' @importFrom rlang .data
 #' @importFrom assertthat assert_that
 #' @importFrom stats end